test/p2p-webcam-app/pyproject.toml
[tool.poetry]
name = "p2p-webcam-app"
version = "0.1.0"
description = ""
authors = ["YeonhyukKim <brankein13@gm.gist.ac.kr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiortc = "^1.11.0"
opencv-python = "^4.11.0.86"
python-socketio = "^5.13.0"
aiohttp = "^3.11.16"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
